requirements.txt
streamlit
pandas
pyyaml
pdfplumber
plotly
numpy
jinja2
requests
altair
python-dateutil
pydeck
rich
markdown-it-py
mdurl
pygments
# Version: v0.3.0
# Add any new dependencies for UI enhancements or sidebar features
# e.g. for advanced metrics or visualization:
# plotly
# streamlit-extras
